+++ database/migrations/create_custom_fields_table.php
@@ -124,7 +124,7 @@ return new class extends Migration
             $table->text('string_value')->nullable();
             $table->longText('text_value')->nullable();
             $table->boolean('boolean_value')->nullable();
-            $table->integer('integer_value')->nullable();
+            $table->bigInteger('integer_value')->nullable();
             $table->double('float_value')->nullable();
             $table->date('date_value')->nullable();
             $table->dateTime('datetime_value')->nullable();